Split matching and pre-matching functions
Matching is used for strict handler matching when selecting the handlers for invocation.

Pre-matching is an optimisation measure — it is used in the early stage of processing to check if it is worth continuing to more heavy computations down the stack. For example, to check if a finalizer should be added/removed.

Some computation-heavy and very variable criteria —essentially, the diffs-matching ones— are excluded from pre-matching, leaving it only to the stable body-matching criteria only.